+++ jgclark.DashboardReact/src/types.js
@@ -205,14 +205,6 @@ export type TSharedSettings = {
   [key: string]: any,
 }
 
-// Feature flags for trying out features without other users seeing them
-// FeatureFlags are stored in the DataStore.settings.featureFlags JSON string
-export type FeatureFlags = {
-  autoRefresh: boolean,
-  overdueProcessing: boolean,
-  metaTooltips: boolean
-}
-
 export type TDropdownItemType = 'switch' | 'input' | 'combo' | 'text' | 'separator' | 'heading' | 'header';
  
 export type TDropdownItem = {